Move the type system over
